fix(client): rename useUser to useCurrentUser.ts
