+++ packages/client/src/hooks/index.ts
@@ -1,5 +1,6 @@
 export { default as useAuth } from './useAuth';
 export { default as useAxios } from './useAxios';
+export { default as useCurrentUser } from './useCurrentUser';
 export { default as useHotkeys } from './useHotkeys';
 export { default as useMediaDevice } from './useMediaDevice';
 export { default as useMediaQuery } from './useMediaQuery';
