+++ packages/client/src/hooks/useCurrentUser.ts
@@ -1,0 +1,7 @@
+import { useContext } from 'react';
+
+import { CurrentUserContext } from '../providers/api/CurrentUserProvider/CurrentUserProvider.context';
+
+const useCurrentUser = () => useContext(CurrentUserContext);
+
+export default useCurrentUser;
