--- packages/client/src/hooks/useUser.ts
@@ -1,7 +1,0 @@
-import { useContext } from 'react';
-
-import { UserContext } from '../providers/api/UserProvider/UserProvider.context';
-
-const useUser = () => useContext(UserContext);
-
-export default useUser;
