+++ packages/client/src/providers/api/CurrentUserProvider/CurrentUserProvider.context.ts
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+import { CurrentUserContextValue } from './CurrentUserProvider.types';
+
+export const CurrentUserContext = createContext<CurrentUserContextValue>([undefined, () => {}]);
