--- packages/client/src/providers/api/UserProvider/UserProvider.context.ts
@@ -1,10 +1,0 @@
-import { createContext } from 'react';
-
-import { UserContextValue } from './UserProvider.types';
-
-export const UserContext = createContext<UserContextValue>({
-  getUser: () => {
-    return undefined;
-  },
-  setUser: () => {},
-});
